--- requirements.txt
@@ -1,6 +1,0 @@
-Flask==3.1.2
-flask-cors==6.0.1
-TTS==0.22.0
-pyyaml==6.0.2
-gunicorn==23.0.0
-flask-openapi3[swagger,redoc,rapidoc,rapipdf,scalar,elements]==4.2.1
